+++ src/MetaheuristicsPlatform/Algorithms/Multimodal/AdaptiveNichingPso/AdaptiveNichingPsoOptimizerReferences.cs
@@ -1,0 +1,14 @@
+using MetaheuristicsPlatform.Classification;
+
+namespace MetaheuristicsPlatform.Algorithms.Multimodal.AdaptiveNichingPso;
+
+public static class AdaptiveNichingPsoOptimizerReferences
+{
+    public static ScientificReference Primary { get; } =
+        new(
+            "Bird & Li",
+            2006,
+            "Adaptively choosing niching parameters in a PSO",
+            "GECCO 2006, 3-10",
+            "10.1145/1143997.1143999");
+}